works

+++ philo/src/utils.c
@@ -5,8 +5,8 @@
 /*                                                    +:+ +:+         +:+     */
 /*   By: jerdos-s <marvin@42lausanne.ch>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/10/24 15:25:47 by jerdos-s          #+#    #+#             */
-/*   Created: 2022/10/24 15:25:47 by jerdos-s         ###   ########.fr       */
+/*   Created: 2022/10/26 15:36:26 by jerdos-s          #+#    #+#             */
+/*   Updated: 2022/10/26 15:36:28 by jerdos-s         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,13 +17,13 @@ unsigned long long	ft_atoi(char *str)
 	unsigned long long	current;
 
 	current = 0;
-	while(*str >= '0' && *str <= '9')
+	while (*str >= '0' && *str <= '9')
 	{
 		current *= 10;
 		current += *str - '0';
 		str++;
 	}
-	return ((int)current);
+	return (current);
 }
 
 unsigned long long	timestamp(void)
